REFACTOR: organize code
