+++ source/berkeley_humanoid_lite/berkeley_humanoid_lite/__init__.py
@@ -1,4 +1,1 @@
 # Copyright (c) 2025, The Berkeley Humanoid Lite Project Developers.
-
-# Register Gym environments.
-from .tasks import *
