+++ source/berkeley_humanoid_lite/berkeley_humanoid_lite/environments/__init__.py
@@ -1,0 +1,1 @@
+from .mujoco import *
